requirements.txt
python-dotenv
langchain-core
langchain-openai
langchain-chroma
langgraph